requirements.txt
alembic==1.4.2
aniso8601==8.0.0
click==7.1.2
coverage==5.2.1
factory-boy==3.0.1
Faker==4.1.2
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
text-unidecode==1.3
Werkzeug==1.0.1
